+++ provider/resource_service.go
@@ -172,6 +172,17 @@ func resourceService() *schema.Resource {
 				Description: "Owner Teams associated with this service",
 			},
 
+			"owners_user_ids": &schema.Schema{
+				Type: schema.TypeList,
+				Elem: &schema.Schema{
+					Type: schema.TypeInt,
+				},
+				Computed:    true,
+				Required:    false,
+				Optional:    true,
+				Description: "Owner Users associated with this service",
+			},
+
 			"slack_channels": &schema.Schema{
 				Type:        schema.TypeList,
 				Computed:    true,
@@ -270,6 +281,9 @@ func resourceServiceCreate(ctx context.Context, d *schema.ResourceData, meta int
 	if value, ok := d.GetOkExists("owners_group_ids"); ok {
 		s.OwnersGroupIds = value.([]interface{})
 	}
+	if value, ok := d.GetOkExists("owners_user_ids"); ok {
+		s.OwnersUserIds = value.([]interface{})
+	}
 	if value, ok := d.GetOkExists("slack_channels"); ok {
 		s.SlackChannels = value.([]interface{})
 	}
@@ -321,6 +335,7 @@ func resourceServiceRead(ctx context.Context, d *schema.ResourceData, meta inter
 	d.Set("environment_ids", item.EnvironmentIds)
 	d.Set("service_ids", item.ServiceIds)
 	d.Set("owners_group_ids", item.OwnersGroupIds)
+	d.Set("owners_user_ids", item.OwnersUserIds)
 	d.Set("slack_channels", item.SlackChannels)
 	d.Set("slack_aliases", item.SlackAliases)
 
@@ -380,6 +395,9 @@ func resourceServiceUpdate(ctx context.Context, d *schema.ResourceData, meta int
 	}
 	if d.HasChange("owners_group_ids") {
 		s.OwnersGroupIds = d.Get("owners_group_ids").([]interface{})
+	}
+	if d.HasChange("owners_user_ids") {
+		s.OwnersUserIds = d.Get("owners_user_ids").([]interface{})
 	}
 	if d.HasChange("slack_channels") {
 		s.SlackChannels = d.Get("slack_channels").([]interface{})